+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Shows missing documents in Emulator UI Firestore viewer.
 - Better supports paths with special characters in Emulator UI Firestore viewer.
+- Better error handling behavior for `firestore:delete` when deleting very large documents.